feat: manage a session's start and staffing from the session page
Issue #1501 retires My Sessions in favour of a calendar Schedule. Its
companion #1074 moved session create into the Program Inventory popup and
edit onto the session page, but left two things behind, and SessionsLive
is their only host anywhere:

  - Start session, the provider's sole :scheduled -> :in_progress trigger.
    ParticipationLive gates its whole attendance UI on :in_progress and
    offers Complete but not Start, so deleting the list as #1501 specifies
    would leave attendance permanently unreachable.
  - The per-session staffing panel from #782/#1413, whose modal has exactly
    one call site.

Both now live on the session page, which is where the Program Inventory
popup already links and where Complete and Edit already are.

The move is not merely capability-preserving. ParticipationLiveHandlers
already states the rule for this surface: a list must send session_id from
the client because it renders one button per row, while a page showing one
session must not, since that hands a tampered event the id the context gate
exists to check (#1373). Start and the staffing trigger therefore read the
id from the mount-time assign, and start_session/2 reloads through its
reload_fn rather than waiting for the {:session_changed, id} broadcast that
a roster page's catch-all handle_info swallows.

Two helpers simplify in the move: the staffing modal reads the program title
off @session.program_name instead of a program_names map this page does not
carry, and session_label/1 drops a get_session/1 round-trip that only existed
because a stream is not enumerable state to search.

SessionsLive keeps Start until #1501 deletes the page; duplicating one
lifecycle handler for a few days is cheaper than rewriting ten assertions in
a file about to be removed. The staffing panel is not duplicated -- five
events and a modal respelled at two surfaces is the shape ADR-0019 rejects.

Refs #1501
