+++ test/klass_hero_web/live/provider/sessions_live_test.exs
@@ -729,174 +729,6 @@ defmodule KlassHeroWeb.Provider.SessionsLiveTest do
     end
   end
 
-  describe "session staffing panel (#782)" do
-    setup :register_and_log_in_provider
-
-    setup %{provider: provider} do
-      program = insert(:program_schema, provider_id: provider.id, title: "Judo")
-
-      session =
-        insert(:program_session_schema,
-          program_id: program.id,
-          session_date: Date.utc_today(),
-          status: :scheduled
-        )
-
-      regular = insert(:staff_member_schema, provider_id: provider.id, first_name: "Ana", last_name: "Stone")
-
-      {:ok, _} =
-        KlassHero.Provider.assign_staff_to_program(%{
-          provider_id: provider.id,
-          program_id: program.id,
-          staff_member_id: regular.id
-        })
-
-      {:ok, program: program, session: session, regular: regular}
-    end
-
-    defp open_panel(view, session) do
-      view |> element("#manage-session-staffing-#{session.id}") |> render_click()
-      view
-    end
-
-    test "opens showing the program's roster and says it is inherited", ctx do
-      {:ok, view, _html} = live(ctx.conn, ~p"/provider/sessions")
-
-      open_panel(view, ctx.session)
-
-      assert has_element?(view, "#session-staffing-modal")
-      assert has_element?(view, "#session-staffing-member-#{ctx.regular.id}")
-      assert render(view) =~ "Using the program"
-      # Nothing to revert to while the roster is still the program's.
-      refute has_element?(view, "#session-staffing-revert-btn")
-    end
-
-    test "adding someone keeps the program's roster and appends to it", ctx do
-      substitute = insert(:staff_member_schema, provider_id: ctx.provider.id, first_name: "Bea", last_name: "Stone")
-
-      {:ok, view, _html} = live(ctx.conn, ~p"/provider/sessions")
-      open_panel(view, ctx.session)
-
-      view
-      |> element("#session-staffing-add-form")
-      |> render_submit(%{"add_staff" => %{"staff_id" => substitute.id}})
-
-      # Both, not one: "Add" adds. The session takes its own roster, but it takes
-      # the program's team with it rather than discarding them.
-      assert has_element?(view, "#session-staffing-member-#{substitute.id}")
-      assert has_element?(view, "#session-staffing-member-#{ctx.regular.id}")
-      assert has_element?(view, "#session-staffing-revert-btn")
-      assert render(view) =~ "Staffed just for this session"
-    end
-
-    test "the picker does not offer people the session already shows", ctx do
-      {:ok, view, _html} = live(ctx.conn, ~p"/provider/sessions")
-      open_panel(view, ctx.session)
-
-      refute has_element?(view, "#session-staffing-add-select option[value='#{ctx.regular.id}']")
-      assert has_element?(view, "#session-staffing-nobody-addable")
-    end
-
-    test "removing works on an inherited roster and leaves the rest", ctx do
-      other = insert(:staff_member_schema, provider_id: ctx.provider.id, first_name: "Cal", last_name: "Stone")
-
-      {:ok, _} =
-        KlassHero.Provider.assign_staff_to_program(%{
-          provider_id: ctx.provider.id,
-          program_id: ctx.program.id,
-          staff_member_id: other.id
-        })
-
-      {:ok, view, _html} = live(ctx.conn, ~p"/provider/sessions")
-      open_panel(view, ctx.session)
-
-      view |> element("#remove-session-staff-#{ctx.regular.id}") |> render_click()
-
-      refute has_element?(view, "#session-staffing-member-#{ctx.regular.id}")
-      assert has_element?(view, "#session-staffing-member-#{other.id}")
-    end
-
-    test "the last member's removal is disabled rather than hidden", ctx do
-      {:ok, view, _html} = live(ctx.conn, ~p"/provider/sessions")
-      open_panel(view, ctx.session)
-
-      assert has_element?(view, "#remove-session-staff-#{ctx.regular.id}[disabled]")
-    end
-
-    test "promote and remove are offered on a roster that is still inherited", ctx do
-      other = insert(:staff_member_schema, provider_id: ctx.provider.id)
-
-      {:ok, _} =
-        KlassHero.Provider.assign_staff_to_program(%{
-          provider_id: ctx.provider.id,
-          program_id: ctx.program.id,
-          staff_member_id: other.id
-        })
-
-      {:ok, view, _html} = live(ctx.conn, ~p"/provider/sessions")
-      open_panel(view, ctx.session)
-
-      refute render(view) =~ "Staffed just for this session"
-      assert has_element?(view, "#promote-session-staff-#{ctx.regular.id}")
-      assert has_element?(view, "#remove-session-staff-#{ctx.regular.id}:not([disabled])")
-    end
-
-    test "reverting returns the session to the program roster", ctx do
-      substitute = insert(:staff_member_schema, provider_id: ctx.provider.id)
-
-      {:ok, _} =
-        KlassHero.Provider.assign_staff_to_session(%{
-          provider_id: ctx.provider.id,
-          session_id: ctx.session.id,
-          staff_member_id: substitute.id
-        })
-
-      {:ok, view, _html} = live(ctx.conn, ~p"/provider/sessions")
-      open_panel(view, ctx.session)
-
-      view |> element("#session-staffing-revert-btn") |> render_click()
-
-      assert has_element?(view, "#session-staffing-member-#{ctx.regular.id}")
-      refute has_element?(view, "#session-staffing-revert-btn")
-    end
-
-    test "promoting a session lead badges them and blocks their removal", ctx do
-      substitute = insert(:staff_member_schema, provider_id: ctx.provider.id)
-
-      {:ok, _} =
-        KlassHero.Provider.assign_staff_to_session(%{
-          provider_id: ctx.provider.id,
-          session_id: ctx.session.id,
-          staff_member_id: substitute.id
-        })
-
-      {:ok, view, _html} = live(ctx.conn, ~p"/provider/sessions")
-      open_panel(view, ctx.session)
-
-      view |> element("#promote-session-staff-#{substitute.id}") |> render_click()
-
-      assert has_element?(view, "#session-staffing-lead-badge-#{substitute.id}")
-      assert has_element?(view, "#remove-session-staff-#{substitute.id}[disabled]")
-    end
-
-    test "a foreign session cannot be opened", ctx do
-      other_provider = insert(:provider_profile_schema)
-      other_program = insert(:program_schema, provider_id: other_provider.id)
-
-      foreign =
-        insert(:program_session_schema, program_id: other_program.id, session_date: Date.utc_today())
-
-      {:ok, view, _html} = live(ctx.conn, ~p"/provider/sessions")
-
-      # The button is not rendered for a session they cannot see, so drive the
-      # event directly — that is the shape a tampered client takes.
-      render_click(view, "manage_session_staffing", %{"id" => foreign.id})
-
-      refute has_element?(view, "#session-staffing-modal")
-      assert_flash(view, :error, "That session could not be found.")
-    end
-  end
-
   # Reads the current :error flash without asserting on it, so two refusals can be
   # compared to each other rather than each to a literal. Same source
   # `assert_flash/3` uses — the socket, not the rendered HTML.
